% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degr.R
\name{new_degr}
\alias{new_degr}
\alias{na_degr}
\alias{format.rastro_degr}
\alias{vec_ptype_abbr.rastro_degr}
\alias{vec_ptype_full.rastro_degr}
\alias{vec_ptype2.rastro_degr}
\alias{vec_ptype2.rastro_degr.default}
\alias{vec_ptype2.rastro_degr.rastro_degr}
\alias{vec_ptype2.rastro_degr.double}
\alias{vec_ptype2.rastro_degr.integer}
\alias{vec_ptype2.integer.rastro_degr}
\alias{vec_ptype2.double.rastro_degr}
\alias{vec_ptype2.rastro_dec.rastro_degr}
\alias{vec_ptype2.rastro_ra.rastro_degr}
\alias{vec_ptype2.rastro_degr.rastro_dec}
\alias{vec_ptype2.rastro_degr.rastro_ra}
\alias{vec_ptype2.rastro_dec.rastro_ra}
\alias{vec_ptype2.rastro_ra.rastro_dec}
\alias{is_degr}
\alias{vec_cast.rastro_degr}
\alias{vec_cast.rastro_degr.default}
\alias{vec_cast.rastro_degr.rastro_degr}
\alias{vec_cast.rastro_degr.double}
\alias{vec_cast.rastro_degr.integer}
\alias{vec_cast.rastro_degr.rastro_dec}
\alias{vec_cast.rastro_degr.rastro_ra}
\alias{vec_cast.double.rastro_degr}
\alias{vec_cast.integer.rastro_degr}
\alias{vec_cast.rastro_dec.rastro_degr}
\alias{vec_cast.rastro_ra.rastro_degr}
\alias{as_degr}
\alias{\%==\%.rastro_degr}
\alias{\%==\%.rastro_degr.default}
\alias{vec_arith.rastro_degr.default}
\alias{vec_arith.rastro_degr.MISSING}
\alias{vec_arith.rastro_degr.numeric}
\alias{vec_arith.numeric.rastro_degr}
\alias{vec_arith.rastro_degr.rastro_degr}
\alias{vec_math.rastro_degr}
\title{Degree}
\usage{
new_degr(deg = double())

na_degr()

\method{format}{rastro_degr}(x, format = "{deg}°", na_string = "NA_rastro_degr_", ...)

vec_ptype_abbr.rastro_degr(x, ...)

vec_ptype_full.rastro_degr(x, ...)

vec_ptype2.rastro_degr(x, y, ...)

\method{vec_ptype2.rastro_degr}{default}(x, y, ..., x_arg = "x", y_arg = "y")

\method{vec_ptype2.rastro_degr}{rastro_degr}(x, y, ...)

\method{vec_ptype2.rastro_degr}{double}(x, y, ...)

\method{vec_ptype2.rastro_degr}{integer}(x, y, ...)

\method{vec_ptype2.integer}{rastro_degr}(x, y, ...)

\method{vec_ptype2.double}{rastro_degr}(x, y, ...)

\method{vec_ptype2.rastro_dec}{rastro_degr}(x, y, ...)

\method{vec_ptype2.rastro_ra}{rastro_degr}(x, y, ...)

\method{vec_ptype2.rastro_degr}{rastro_dec}(x, y, ...)

\method{vec_ptype2.rastro_degr}{rastro_ra}(x, y, ...)

\method{vec_ptype2.rastro_dec}{rastro_ra}(x, y, ...)

\method{vec_ptype2.rastro_ra}{rastro_dec}(x, y, ...)

is_degr(x)

vec_cast.rastro_degr(x, to, ...)

\method{vec_cast.rastro_degr}{default}(x, to, ...)

\method{vec_cast.rastro_degr}{rastro_degr}(x, to, ...)

\method{vec_cast.rastro_degr}{double}(x, to, ...)

\method{vec_cast.rastro_degr}{integer}(x, to, ...)

\method{vec_cast.rastro_degr}{rastro_dec}(x, to, ...)

\method{vec_cast.rastro_degr}{rastro_ra}(x, to, ...)

\method{vec_cast.double}{rastro_degr}(x, to, ...)

\method{vec_cast.integer}{rastro_degr}(x, to, ...)

\method{vec_cast.rastro_dec}{rastro_degr}(x, to, ...)

\method{vec_cast.rastro_ra}{rastro_degr}(x, to, ...)

as_degr(x)

`\%==\%.rastro_degr`(x, y)

\method{\%==\%.rastro_degr}{default}(x, y)

\method{vec_arith.rastro_degr}{default}(op, x, y, ...)

\method{vec_arith.rastro_degr}{MISSING}(op, x, y, ...)

\method{vec_arith.rastro_degr}{numeric}(op, x, y, ...)

\method{vec_arith.numeric}{rastro_degr}(op, x, y, ...)

\method{vec_arith.rastro_degr}{rastro_degr}(op, x, y, ...)

vec_math.rastro_degr(.fn, .x, ...)
}
\arguments{
\item{deg}{Degrees to store.}

\item{x, y}{\code{vec_ptype2*} arguments.}

\item{format, na_string}{\code{glue} flromat strings (support interpolation).}

\item{...}{Additional parameters.}

\item{x_arg, y_arg}{\code{vec_ptype2*} error message variable names.}

\item{to}{\code{vec_cast} argument.}

\item{op, .fn}{Arithmetic functions/operators.}

\item{.x}{\code{vec_arith*} argument.}
}
\description{
Degree
}
