% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{new_flux}
\alias{new_flux}
\alias{na_flux}
\alias{format.rastro_flux}
\alias{obj_print_footer.rastro_flux}
\alias{vec_ptype_abbr.rastro_flux}
\alias{vec_ptype_full.rastro_flux}
\alias{vec_ptype2.rastro_flux}
\alias{vec_ptype2.rastro_flux.default}
\alias{vec_ptype2.rastro_flux.rastro_flux}
\alias{vec_ptype2.rastro_flux.double}
\alias{vec_ptype2.rastro_flux.integer}
\alias{vec_ptype2.integer.rastro_flux}
\alias{vec_ptype2.double.rastro_flux}
\alias{is_flux}
\alias{is.na.rastro_flux}
\alias{vec_cast.rastro_flux}
\alias{vec_cast.rastro_flux.default}
\alias{vec_cast.rastro_flux.rastro_flux}
\alias{vec_cast.rastro_flux.integer}
\alias{vec_cast.rastro_flux.double}
\alias{vec_cast.double.rastro_flux}
\alias{vec_cast.integer.rastro_flux}
\alias{as_flux}
\alias{\%==\%.rastro_flux}
\alias{\%==\%.rastro_flux.default}
\alias{vec_arith.rastro_flux}
\alias{vec_arith.rastro_flux.default}
\alias{vec_arith.rastro_flux.MISSING}
\alias{vec_arith.rastro_flux.numeric}
\alias{vec_arith.numeric.rastro_flux}
\alias{vec_arith.rastro_flux.rastro_flux}
\alias{vec_math.rastro_flux}
\title{Flux}
\usage{
new_flux(flux = double(), filter = NA_character_, unit = NA_character_)

na_flux()

\method{format}{rastro_flux}(x, format = "{flux:\%.3e}", na_string = "NA_rastro_flux_", ...)

obj_print_footer.rastro_flux(x, ...)

vec_ptype_abbr.rastro_flux(x, ...)

vec_ptype_full.rastro_flux(x, ...)

vec_ptype2.rastro_flux(x, y, ...)

\method{vec_ptype2.rastro_flux}{default}(x, y, ..., x_arg = "x", y_arg = "y")

\method{vec_ptype2.rastro_flux}{rastro_flux}(x, y, ..., x_arg = "x", y_arg = "y")

\method{vec_ptype2.rastro_flux}{double}(x, y, ...)

\method{vec_ptype2.rastro_flux}{integer}(x, y, ...)

\method{vec_ptype2.integer}{rastro_flux}(x, y, ...)

\method{vec_ptype2.double}{rastro_flux}(x, y, ...)

is_flux(x, filter = NA_character_, unit = NA_real_)

\method{is.na}{rastro_flux}(x)

vec_cast.rastro_flux(x, to, ..., x_arg = "x", to_arg = "to")

\method{vec_cast.rastro_flux}{default}(x, to, ..., x_arg = "x", to_arg = "to")

\method{vec_cast.rastro_flux}{rastro_flux}(x, to, ..., x_arg = "x", to_arg = "to")

\method{vec_cast.rastro_flux}{integer}(x, to, ...)

\method{vec_cast.rastro_flux}{double}(x, to, ...)

\method{vec_cast.double}{rastro_flux}(x, to, ...)

\method{vec_cast.integer}{rastro_flux}(x, to, ...)

as_flux(x, filter = NA_character_, unit = NA_real_, ...)

`\%==\%.rastro_flux`(x, y)

\method{\%==\%.rastro_flux}{default}(x, y)

vec_arith.rastro_flux(op, x, y, ...)

\method{vec_arith.rastro_flux}{default}(op, x, y, ...)

\method{vec_arith.rastro_flux}{MISSING}(op, x, y, ...)

\method{vec_arith.rastro_flux}{numeric}(op, x, y, ...)

\method{vec_arith.numeric}{rastro_flux}(op, x, y, ...)

\method{vec_arith.rastro_flux}{rastro_flux}(op, x, y, ...)

vec_math.rastro_flux(.fn, .x, ...)
}
\arguments{
\item{flux, new_flux}{Flux values.}

\item{filter}{An optional filter.}

\item{unit}{Am optional measurement units.}

\item{x, y}{\code{vec_ptype2*} arguments.}

\item{format, na_string}{\code{glue} flromat strings (support interpolation).}

\item{...}{Additional parameters.}

\item{x_arg, y_arg, to_arg}{\code{vec_ptype2*} and \code{vec_cast*} error message variable names.}

\item{to}{\code{vec_cast} argument.}

\item{op, .fn}{Arithmetic functions/operators.}

\item{.x}{\code{vec_arith*} argument.}
}
\description{
Flux
}
