% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mag.R
\name{new_mag}
\alias{new_mag}
\alias{na_mag}
\alias{format.rastro_mag}
\alias{obj_print_footer.rastro_mag}
\alias{vec_ptype_abbr.rastro_mag}
\alias{vec_ptype_full.rastro_mag}
\alias{vec_ptype2.rastro_mag}
\alias{vec_ptype2.rastro_mag.default}
\alias{vec_ptype2.rastro_mag.rastro_mag}
\alias{vec_ptype2.rastro_mag.double}
\alias{vec_ptype2.rastro_mag.integer}
\alias{vec_ptype2.integer.rastro_mag}
\alias{vec_ptype2.double.rastro_mag}
\alias{is_mag}
\alias{vec_cast.rastro_mag}
\alias{vec_cast.rastro_mag.default}
\alias{vec_cast.rastro_mag.rastro_mag}
\alias{vec_cast.rastro_mag.integer}
\alias{vec_cast.integer.rastro_mag}
\alias{vec_cast.rastro_mag.double}
\alias{vec_cast.double.rastro_mag}
\alias{as_mag}
\alias{\%==\%.rastro_mag}
\alias{\%==\%.rastro_mag.default}
\alias{vec_arith.rastro_mag}
\alias{vec_arith.rastro_mag.default}
\alias{vec_arith.rastro_mag.MISSING}
\alias{vec_arith.rastro_mag.numeric}
\alias{vec_arith.numeric.rastro_mag}
\alias{vec_arith.rastro_mag.rastro_mag}
\alias{vec_math.rastro_mag}
\alias{vec_cast.rastro_flux.rastro_mag}
\title{Magnitude}
\usage{
new_mag(m = double(), filter = NA_character_, zero_flux = na_flux())

na_mag()

\method{format}{rastro_mag}(x, format = "{mag:\%.3f}", na_string = "NA_rastro_mag_", ...)

obj_print_footer.rastro_mag(x, ...)

\method{vec_ptype_abbr}{rastro_mag}(x, ...)

\method{vec_ptype_full}{rastro_mag}(x, ...)

vec_ptype2.rastro_mag(x, y, ...)

\method{vec_ptype2.rastro_mag}{default}(x, y, ..., x_arg = "x", y_arg = "y")

\method{vec_ptype2.rastro_mag}{rastro_mag}(x, y, ..., x_arg = "x", y_arg = "y")

\method{vec_ptype2.rastro_mag}{double}(x, y, ...)

\method{vec_ptype2.rastro_mag}{integer}(x, y, ...)

\method{vec_ptype2.integer}{rastro_mag}(x, y, ...)

\method{vec_ptype2.double}{rastro_mag}(x, y, ...)

is_mag(x, filter = NA_character_, zero_flux = NA_real_)

vec_cast.rastro_mag(x, to, ..., x_arg = "x", to_arg = "to")

\method{vec_cast.rastro_mag}{default}(x, to, ..., x_arg = "x", to_arg = "to")

\method{vec_cast.rastro_mag}{rastro_mag}(x, to, ..., x_arg = "x", to_arg = "to")

\method{vec_cast.rastro_mag}{integer}(x, to, ...)

\method{vec_cast.integer}{rastro_mag}(x, to, ...)

\method{vec_cast.rastro_mag}{double}(x, to, ...)

\method{vec_cast.double}{rastro_mag}(x, to, ...)

as_mag(x, filter = NA_character_, zero_flux = NA_real_, ...)

`\%==\%.rastro_mag`(x, y)

\method{\%==\%.rastro_mag}{default}(x, y)

vec_arith.rastro_mag(op, x, y, ...)

\method{vec_arith.rastro_mag}{default}(op, x, y, ...)

\method{vec_arith.rastro_mag}{MISSING}(op, x, y, ...)

\method{vec_arith.rastro_mag}{numeric}(op, x, y, ...)

\method{vec_arith.numeric}{rastro_mag}(op, x, y, ...)

\method{vec_arith.rastro_mag}{rastro_mag}(op, x, y, ...)

vec_math.rastro_mag(.fn, .x, ...)

\method{vec_cast.rastro_flux}{rastro_mag}(x, to, ..., x_arg = "x", to_arg = "to")
}
\arguments{
\item{m, new_mag}{Magnitude values.}

\item{filter}{An optional filter.}

\item{zero_flux}{An optional \code{rastro::rastro_flux} zero flux.}

\item{x, y}{\code{vec_ptype2*} arguments.}

\item{format, na_string}{\code{glue} flromat strings (support interpolation).}

\item{...}{Additional parameters.}

\item{x_arg, y_arg, to_arg}{\code{vec_ptype2*} and \code{vec_cast*} error message variable names.}

\item{to}{\code{vec_cast} argument.}

\item{op, .fn}{Arithmetic functions/operators.}

\item{.x}{\code{vec_arith*} argument.}
}
\description{
Magnitude
}
