% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_df_interop.R
\name{vec_ptype2.tbl_df}
\alias{vec_ptype2.tbl_df}
\alias{vec_ptype2.tbl_df.default}
\alias{vec_ptype2.tbl_df.data.frame}
\alias{vec_ptype2.tbl_df.tbl_df}
\alias{vec_cast.tbl_df}
\alias{vec_cast.tbl_df.default}
\alias{vec_cast.tbl_df.data.frame}
\alias{vec_cast.tbl_df.tbl_df}
\title{Tibble methods for comaptibility}
\usage{
vec_ptype2.tbl_df(x, y, ...)

\method{vec_ptype2.tbl_df}{default}(x, y, ...)

\method{vec_ptype2.tbl_df}{data.frame}(x, y, ...)

\method{vec_ptype2.tbl_df}{tbl_df}(x, y, ...)

vec_cast.tbl_df(x, to, ...)

\method{vec_cast.tbl_df}{default}(x, to, ...)

\method{vec_cast.tbl_df}{data.frame}(x, to, ...)

\method{vec_cast.tbl_df}{tbl_df}(x, to, ...)
}
\arguments{
\item{x, y}{Vectors to test type/cast.}

\item{...}{Additional parameters}

\item{to}{To what type to cast.}
}
\description{
Tibble methods for comaptibility
}
